Rename the app package to artalk and make it installable
"app" is not an importable identity outside this repository and
collides with other projects' package names. Rename the package to
artalk, update the imports, and add minimal pyproject metadata so the
repository can be installed with pip (including editable installs) and
imported as `artalk` by external integrations. Pure rename otherwise:
no content changes.
