+++ artalk/modules/bitwise_vae.py
