+++ artalk/modules/style_encoder.py
