+++ pyproject.toml
@@ -1,0 +1,13 @@
+[build-system]
+requires = ["setuptools>=68"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "artalk"
+version = "0.1.0"
+description = "ARTalk: speech-driven 3D head animation"
+readme = "README.md"
+requires-python = ">=3.10"
+
+[tool.setuptools.packages.find]
+include = ["artalk*"]
